refactor(/is): general update.
Update:
simplify the default `payload` object that is being passed to the callback.
parameter `payload` of the provided function `callback` is of a type `payload` of the main function.
`payload` parameter is now a `CallbackPayload` type that includes `Payload` instead just a `Payload`.
`jsdoc` description.
parameters `min` and `max` are now `length` and `range` to give a possibility to check only one of them.
remove unnecessary checking especially for the method `hasOwnProperty()`, `in` operator.
`isObjectKeyIn()` and `isObjectKey()`  functions can now check only one key. To check more use `isObjectKeys()` and `isObjectKeysIn()`.
`isObjectKeys()` works differently, now it searches for every key. The previous functionality provides `isObjectSomeKeys()`.
`isPrimitive()` and `isType()` function does not check anymore the `type` argument and the default `switch` returns `false`.
`isStringLength()` change parameters `min` and `max` to an object `length` to check one of them. Simplify the way it checks.
`isNumberBetween()` change parameters `min` and `max` to an object `range` to check one of them. Simplify the way it checks.

Add:
type of the `value` to the callback function.
the `value` that has been checked to the callback function.
additional properties from the main function if exists e.g `constructor`, `type`, `key`, `keys`, `min`, `max`.